Add the auth unit-tests to SConscript and .pbxproj
Review URL: http://codereview.chromium.org/4315

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@2660 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/http/http_auth_handler_basic_unittest.cc
@@ -19,7 +19,7 @@ TEST(HttpAuthHandlerBasicTest, GenerateCredentials) {
     // Empty password
     { L"anon", L"", "Basic YW5vbjo=" },
   };
-  for (int i = 0; i < ARRAYSIZE_UNSAFE(tests); ++i) {
+  for (size_t i = 0; i < ARRAYSIZE_UNSAFE(tests); ++i) {
     std::string challenge = "Basic realm=\"Atlantis\"";
     HttpAuthHandlerBasic basic;
     basic.InitFromChallenge(challenge.begin(), challenge.end(),
